Added error pages from start bootstrap.
Added a generic error page and dedicated 401 and 404 pages. The generic ones redirects to the dedicated 401 and 404 pages.

Implemented a BaseErrorModel that all error pages use to obtain the request id.

Added UseDeveloperExceptionPage when in development mode.

Added UseStatusCodePagesWithReExecute in non-development mode so that the status code is passed to the Error page.